Adding logging functionality

+++ Projects/Sorting/Sorting/HeapSort.h
@@ -4,27 +4,27 @@
 #include <iostream>
 #include <fstream>
 #include <string>
-
+#include "Logger.h"
 class CHeapSort
 {
 private:
 	std::string m_sFileToSort;
-	int* m_pElement;
-	int m_nindex;
-	int m_nNumberOfElements;
+	__int64* m_pElement;
+	__int64 m_nindex;
+	__int64 m_nNumberOfElements;
 public:
 
 	CHeapSort();
-	CHeapSort(int n);
+	CHeapSort(__int64 n);
 	~CHeapSort();
 
 	void Sort();
 	void FileToSort(std::string filepath);
 	void Parse();
 	void WriteSortedListToFile(std::string filepath);
-	void swim(int k);
-	void insert(int value);
-	void sink(int k);
+	void swim(__int64 k);
+	void insert(__int64 value);
+	void sink(__int64 k);
 	void deleteMAX();
 
 };
